requirements.txt
django==1.9.0
django-crispy-forms>=1.6.0
django-import-export>=0.5.1
django-reversion>=2.0.0
django-formtools==1.0
future==0.15.2
httplib2==0.9.2
six==1.10.0
django-allauth==0.32.0
gitpython==2.1.10
pymysql==0.8.1
python-jenkins==1.4.0
django-apscheduler==0.2.4